fix: no database connection available for a query (#17)
* chore: general linting tasks and code cleaning

for some reason the src/routers is still available in the version
of the develop branch. eslint caught this as a problem and so it has been
deleted to fix. the routers are now part of the middleware folder.

* fix: no database connection available for a query

the model instances were not properly imported, resulting in the db not
being properly initialized with the instance of the models, and the
no database connection available being thrown.

this commit fixes this error.