fix: email regex
Change from (/^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/
to (/^[\w-.]+@([\w-]+\.)+[A-Za-z]{2,}$/i

The change limits the domain to only alphabets and removes the imposed maxlength to the domain. Domain nowadays have more than 4
characters and is likely to increase in the future

Added the insensitive flag (`i`)so as to accept emails written in
big caps. The email will be sanitized by the sanitation function.